añadiendo repos

+++ src/main/java/com/example/demo/model/PedidoTable.java
@@ -1,5 +1,7 @@
 package com.example.demo.model;
 
+import java.math.BigDecimal;
+
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
@@ -15,6 +17,6 @@ import lombok.NoArgsConstructor;
 public class PedidoTable {
     private @Column("PEDIDO_ID") @Id Long id;
     private @Column("MESA") Long mesa;
-    private @Column("PRECIO") Double precio;
+    private @Column("PRECIO") BigDecimal precio;
     private @Column("ESTADO") Long estado;
 }
